+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Let `resolveStateDirPath` return the correct state path when using the `ifState` environment variable to set the state directory.
+
 ## [1.3.2] - 2024-05-15
 
 ### Fixed